requirements.txt
streamlit
pandas
numpy
scikit-learn
openpyxl
matplotlib
python-dateutil